Avoid numpy warning
FutureWarning: Conversion of the second argument of issubdtype from `int` to `np.signedinteger` is deprecated. In future, it will be treated as `np.int64 == np.dtype(int).type`.

+++ guidedlda/utils.py
@@ -52,7 +52,7 @@ def matrix_to_lists(doc_word):
     except AttributeError:
         sparse = False
 
-    if sparse and not np.issubdtype(doc_word.dtype, int):
+    if sparse and not np.issubdtype(doc_word.dtype, np.int64):
         raise ValueError("expected sparse matrix with integer values, found float values")
 
     ii, jj = np.nonzero(doc_word)
